main.basic.js 파일에 있는 함수들을 별도 파일로 빼는 코드 분리 작업
