+++ src/basic/src/updateBonusPoints.js
@@ -1,0 +1,81 @@
+import { PRODUCT_1, PRODUCT_2, PRODUCT_3 } from '../constants.js';
+
+export function onUpdateBonusPoints({ cartDisplay, totalAmt, prodList, itemCnt }) {
+  if (cartDisplay.children.length === 0) {
+    document.getElementById('loyalty-points').style.display = 'none';
+    return;
+  }
+  const basePoints = Math.floor(totalAmt / 1000);
+  let finalPoints = 0;
+  const pointsDetail = [];
+  if (basePoints > 0) {
+    finalPoints = basePoints;
+    pointsDetail.push('기본: ' + basePoints + 'p');
+  }
+  if (new Date().getDay() === 2) {
+    if (basePoints > 0) {
+      finalPoints = basePoints * 2;
+      pointsDetail.push('화요일 2배');
+    }
+  }
+  let hasKeyboard = false;
+  let hasMouse = false;
+  let hasMonitorArm = false;
+  const nodes = cartDisplay.children;
+  for (const node of nodes) {
+    let product = null;
+    for (let pIdx = 0; pIdx < prodList.length; pIdx++) {
+      if (prodList[pIdx].id === node.id) {
+        product = prodList[pIdx];
+        break;
+      }
+    }
+    if (!product) continue;
+    if (product.id === PRODUCT_1) {
+      hasKeyboard = true;
+    } else if (product.id === PRODUCT_2) {
+      hasMouse = true;
+    } else if (product.id === PRODUCT_3) {
+      hasMonitorArm = true;
+    }
+  }
+  if (hasKeyboard && hasMouse) {
+    finalPoints = finalPoints + 50;
+    pointsDetail.push('키보드+마우스 세트 +50p');
+  }
+  if (hasKeyboard && hasMouse && hasMonitorArm) {
+    finalPoints = finalPoints + 100;
+    pointsDetail.push('풀세트 구매 +100p');
+  }
+  if (itemCnt >= 30) {
+    finalPoints = finalPoints + 100;
+    pointsDetail.push('대량구매(30개+) +100p');
+  } else {
+    if (itemCnt >= 20) {
+      finalPoints = finalPoints + 50;
+      pointsDetail.push('대량구매(20개+) +50p');
+    } else {
+      if (itemCnt >= 10) {
+        finalPoints = finalPoints + 20;
+        pointsDetail.push('대량구매(10개+) +20p');
+      }
+    }
+  }
+  const bonusPts = finalPoints;
+  const ptsTag = document.getElementById('loyalty-points');
+  if (ptsTag) {
+    if (bonusPts > 0) {
+      ptsTag.innerHTML =
+        '<div>적립 포인트: <span class="font-bold">' +
+        bonusPts +
+        'p</span></div>' +
+        '<div class="text-2xs opacity-70 mt-1">' +
+        pointsDetail.join(', ') +
+        '</div>';
+      ptsTag.style.display = 'block';
+    } else {
+      ptsTag.textContent = '적립 포인트: 0p';
+      ptsTag.style.display = 'block';
+    }
+  }
+}
