+++ src/basic/src/updatePricesInCart.js
@@ -1,0 +1,46 @@
+export function onUpdatePricesInCart({ cartDisplay, productList, handleCalculateCartStuff }) {
+  const cartItems = cartDisplay.children;
+  for (let i = 0; i < cartItems.length; i++) {
+    const itemId = cartItems[i].id;
+    let product = null;
+    for (let productIdx = 0; productIdx < productList.length; productIdx++) {
+      if (productList[productIdx].id === itemId) {
+        product = productList[productIdx];
+        break;
+      }
+    }
+    if (product) {
+      const priceDiv = cartItems[i].querySelector('.text-lg');
+      const nameDiv = cartItems[i].querySelector('h3');
+      if (product.onSale && product.suggestSale) {
+        priceDiv.innerHTML =
+          '<span class="line-through text-gray-400">₩' +
+          product.originalVal.toLocaleString() +
+          '</span> <span class="text-purple-600">₩' +
+          product.val.toLocaleString() +
+          '</span>';
+        nameDiv.textContent = '⚡💝' + product.name;
+      } else if (product.onSale) {
+        priceDiv.innerHTML =
+          '<span class="line-through text-gray-400">₩' +
+          product.originalVal.toLocaleString() +
+          '</span> <span class="text-red-500">₩' +
+          product.val.toLocaleString() +
+          '</span>';
+        nameDiv.textContent = '⚡' + product.name;
+      } else if (product.suggestSale) {
+        priceDiv.innerHTML =
+          '<span class="line-through text-gray-400">₩' +
+          product.originalVal.toLocaleString() +
+          '</span> <span class="text-blue-500">₩' +
+          product.val.toLocaleString() +
+          '</span>';
+        nameDiv.textContent = '💝' + product.name;
+      } else {
+        priceDiv.textContent = '₩' + product.val.toLocaleString();
+        nameDiv.textContent = product.name;
+      }
+    }
+  }
+  handleCalculateCartStuff();
+}
