+++ src/basic/src/updateStockInfo.js
@@ -1,0 +1,14 @@
+export function onUpdateStockInfo({ productList, stockInfo }) {
+  let infoMsg = '';
+
+  productList.forEach(function (item) {
+    if (item.q < 5) {
+      if (item.q > 0) {
+        infoMsg = infoMsg + item.name + ': 재고 부족 (' + item.q + '개 남음)\n';
+      } else {
+        infoMsg = infoMsg + item.name + ': 품절\n';
+      }
+    }
+  });
+  stockInfo.textContent = infoMsg;
+}
